Begin makeover of first site I ever did
This site was built as a final project for one of my classes. I thought
it would be fun to do a makeover since my HTML and CSS skills have
matured in the past three years. I would like to make the site
responsive, as that was not part of the assignment originally, and I
would also like to improve the overall design aesthetic.